Stage 2.3 (section-align Events): rename repository + folder + namespace
- IEventGuideRepository → IEventRepository (interface)
- EventGuideRepository → EventRepository (sealed impl)
- src/Humans.Infrastructure/Repositories/EventGuide/ →
  src/Humans.Infrastructure/Repositories/Events/
- namespace Humans.Infrastructure.Repositories.EventGuide →
  Humans.Infrastructure.Repositories.Events

Word-boundary rename touches type/interface/namespace references only;
test class names (EventGuideRepositoryTests, FakeEventGuideRepository,
arch-test method names like IEventGuideRepository_LivesIn...) use `_`
separators so they don't have a word boundary and get cleaned up in
Stage 2.8 along with the test folder rearrangement.

Baseline path updated in DisplaySortInControllers.baseline.txt
(EventGuide/EventGuideRepository.cs → Events/EventRepository.cs) — same
17 OrderBy/ThenBy entries, new relative path.
