+++ tests/Humans.Application.Tests/Architecture/Baselines/DisplaySortInControllers.baseline.txt
@@ -68,23 +68,23 @@ src/Humans.Infrastructure/Repositories/Consent/ConsentRepository.cs:OrderByDesce
 src/Humans.Infrastructure/Repositories/Consent/ConsentRepository.cs:OrderByDescending#2
 src/Humans.Infrastructure/Repositories/Consent/ConsentRepository.cs:OrderByDescending#3
 src/Humans.Infrastructure/Repositories/Email/EmailOutboxRepository.cs:OrderBy#1
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#1
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#2
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#3
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#4
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#5
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#6
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#7
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#8
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderBy#9
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderByDescending#1
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderByDescending#2
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderByDescending#3
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:OrderByDescending#4
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:ThenBy#1
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:ThenBy#2
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:ThenBy#3
-src/Humans.Infrastructure/Repositories/EventGuide/EventGuideRepository.cs:ThenBy#4
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#1
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#2
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#3
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#4
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#5
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#6
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#7
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#8
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderBy#9
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderByDescending#1
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderByDescending#2
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderByDescending#3
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:OrderByDescending#4
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:ThenBy#1
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:ThenBy#2
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:ThenBy#3
+src/Humans.Infrastructure/Repositories/Events/EventRepository.cs:ThenBy#4
 src/Humans.Infrastructure/Repositories/Feedback/FeedbackRepository.cs:OrderBy#1
 src/Humans.Infrastructure/Repositories/GoogleIntegration/GoogleResourceRepository.cs:OrderBy#1
 src/Humans.Infrastructure/Repositories/GoogleIntegration/GoogleResourceRepository.cs:OrderBy#2
